+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-10-20
+
+### Changed
+
+- "(re)start" button is in footer
+
+### Fixed
+
+- session log is displayed after finish
+
 ## [0.2.2] - 2023-10-16
 
 ### Changed
@@ -39,8 +49,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - autofocus on input after page is loaded
 - score with right answers and total count of problems
 
-[Unreleased]: https://github.com/vikian050194/abacus/compare/v0.2.2...HEAD
-[0.2.2]: https://github.com/vikian050194/abacus/compare/v0.2.0...v0.2.2
+[Unreleased]: https://github.com/vikian050194/abacus/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/vikian050194/abacus/compare/v0.2.2...v0.3.0
+[0.2.2]: https://github.com/vikian050194/abacus/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/vikian050194/abacus/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/vikian050194/abacus/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/vikian050194/abacus/releases/tag/v0.1.0